+Participant data reading
- Minor changes to PacketModel to read the correct length of bytes
- Added new class Participant
- Added new field mNumActiveCars to ParticipantPacketModel
- Moved Participant data from ParticipantPacketModel to Participant
- Added method Participant.computeName to compute the name
- +javaDoc Participant (not complete)
